+++ test/extended_bloc_test.dart
@@ -1,18 +1,10 @@
-//import 'package:extended_bloc/extended_bloc.dart';
 import 'package:test/test.dart';
+import 'src/enum_bloc_test.dart' as enum_bloc;
+import 'src/router_bloc_test.dart' as router_bloc;
 
 void main() {
-  group('Placeholder group', () {
-    setUp(() {});
-
-    tearDown(() {});
-
-    setUpAll(() {});
-
-    tearDownAll(() {});
-
-    test('Placeholder test', () {
-      expect(true, isTrue);
-    });
+  group('extended_bloc', () {
+    enum_bloc.main();
+    router_bloc.main();
   });
 }