+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## [Unreleased]
 ## 0.0.1-dev - 2020-08-30
 ### Added
-- initial version
+- RouterBloc<Event, State>
+- EnumBloc<Event, State>
 
 [Unreleased]: